Move thread_sleep_scheduler and thread_shutdown_scheduler into thread_worker.h
