+++ c/parallel/thread_scheduler.h
@@ -17,8 +17,9 @@
 #define dt_PARALLEL_THREAD_SCHEDULER_h
 #include <atomic>      // std::atomic
 #include <memory>      // std::unique_ptr
+#include <mutex>       // std::mutex
 #include <vector>      // std::vector
-#include "parallel/thread_worker.h"
+#include "parallel/thread_task.h"
 #include "utils/function.h"
 #include "utils/macros.h"
 namespace dt {
@@ -65,100 +66,6 @@ class thread_scheduler {
     // blocking. The default implementation does nothing (all scheduled tasks
     // continue being executed), which is allowed but sub-optimal.
     virtual void abort_execution();
-};
-
-
-
-//------------------------------------------------------------------------------
-// thread sleep scheduler
-//------------------------------------------------------------------------------
-
-/**
- * This class handles putting to sleep/awaking of workers in a thread pool.
- * A single instance of this class exists in `thread_pool`.
- *
- * Initially all workers in a thread pool are in the "idle" state, running the
- * sleep task returned by this scheduler. This sleep task is `tsleep[0]`, and
- * it contains a mutex, and a condition variable. In this state the workers are
- * simply waiting, though they may occasionally be woken by the operating system
- * to check whether `tsleep[0].next_scheduler` became non-null.
- *
- * More precisely, a thread is considered to be asleep if its scheduler is this
- * class, and if the thread already requested a sleep task from this scheduler
- * and started executing that sleep task.
- *
- * When master thread calls `awaken` (and only the master thread is allowed to
- * do so), we do the following:
- *   - lock `tsleep[0].mutex` (at this point no thread can awaken, even
- *     spuriously, because they would need to acquire lock on the same mutex as
- *     they wake up);
- *   - set `tsleep[0].next_scheduler` to the job that needs to be executed;
- *   - set `tsleep[1].next_scheduler` to nullptr;
- *   - change `index` from 0 to 1;
- *   - unlock the mutex and notify all threads waiting on `tsleep[0].alarm`.
- *
- * As the threads awaken, they check their task's `next_scheduler` property, see
- * that it is now not-null, they will switch to that scheduler and finish their
- * current sleep task. Note that it may take some time for OS to notify and
- * awaken all the threads; some threads may already finish their new task by the
- * time the last thread in the team gets up.
- *
- * When a thread's queue is exhausted and there are no more tasks to do, that
- * worker receives a `nullptr` from `get_next_task()`. At this moment the
- * worker switches back to `thread_sleep_scheduler`, and requests a task. The
- * thread sleep scheduler will now return `tsleep[1]`, which has its own mutex
- * and a condition variable, and its `.next_scheduler` is null, indicating the
- * sleeping state. This will allow the thread to go safely to sleep, while other
- * threads might still be waking up from the initial sleep.
- *
- * The master thread that called `awaken(job)` will then call `job.join()`,
- * and it is the responsibility of thread_scheduler `job` to wait until all
- * threads have finished execution and were put back to sleep. Thus, the master
- * thread ensures that all threads are sleeping again before the next call to
- * `awaken`.
- */
-class thread_sleep_scheduler : public thread_scheduler {
-  private:
-    static constexpr size_t N_SLEEP_TASKS = 2;
-    thread_sleep_task tsleep[N_SLEEP_TASKS];
-    // Index within array `tsleep` indicating which sleep task is "current";
-    // where "current" means that all sleeping threads are waiting on the
-    // condition variable within that task, and its `n_sleeping_threads` holds
-    // the total count of workers that are currently sleeping.
-    size_t index = 0;
-
-  public:
-    thread_task* get_next_task(size_t thread_index) override;
-
-    void awaken(thread_scheduler* next);
-
-    // Called from the master thread, this function will block until all the
-    // work is finished and all worker threads have been put to sleep. If there
-    // was an exception during execution of any of the tasks, this exception
-    // will be rethrown here (but only after all workers were put to sleep).
-    void join(size_t nthreads);
-
-  private:
-    friend class thread_shutdown_scheduler;
-    void pretend_thread_went_to_sleep();
-};
-
-
-
-
-//------------------------------------------------------------------------------
-// thread shutdown scheduler
-//------------------------------------------------------------------------------
-
-class thread_shutdown_scheduler : public thread_scheduler {
-  private:
-    size_t n_threads_to_keep;
-    thread_sleep_scheduler* sleep_scheduler;
-    shutdown_thread_task shutdown;
-
-  public:
-    thread_shutdown_scheduler(size_t nnew, thread_sleep_scheduler*);
-    thread_task* get_next_task(size_t thread_index) override;
 };
